+++ hw09_struct_validator/in.go
@@ -1,0 +1,22 @@
+package hw09_struct_validator
+
+import "fmt"
+
+type inValidator struct {
+	Items []string
+}
+
+func (iv inValidator) validate(field string, value interface{}) *ValidationError {
+	str := fmt.Sprintf("%v", value)
+
+	for _, item := range iv.Items {
+		if str == item {
+			return nil
+		}
+	}
+
+	return &ValidationError{
+		Field: field,
+		Err:   ErrValueNotIn,
+	}
+}
